+++ .travis.yml
@@ -15,6 +15,7 @@ matrix:
         - QMAKESPEC=linux-g++
         - EVAL="CC=gcc-7 && CXX=g++-7"
         - CFLAGS="-Os"
+        - LDFLAGS="-Wl,--no-undefined -lm"
         - QMAKEFLAGS="-config release"
     - os: linux
       addons:
@@ -29,9 +30,23 @@ matrix:
         - EVAL="CC=clang-5.0 && CXX=clang++"
         - PATH=$PATH:/usr/local/clang/bin
         - CFLAGS="-Oz"
+        - LDFLAGS="-Wl,--no-undefined -lm"
         - QMAKEFLAGS="-config release"
         - MAKEFLAGS=-s
         - TESTARGS=-silent
+    - os: linux
+      env:
+        - QMAKESPEC=linux-gcc-freestanding
+        - EVAL="CXX=false"
+        - CFLAGS="-ffreestanding -Os"
+        - LDFLAGS="-Wl,--no-undefined -lm"
+    - os: linux
+      env:
+        - QMAKESPEC=linux-gcc-no-math
+        - EVAL="CXX=false && touch src/math.h src/float.h"
+        - CFLAGS="-ffreestanding -DCBOR_NO_FLOATING_POINT -Os"
+        - LDFLAGS="-Wl,--no-undefined"
+        - LDLIBS=""
     - os: osx
       env:
         - QMAKESPEC=macx-clang
@@ -57,10 +72,12 @@ script:
   - make -s clean
   - make -k
         CFLAGS="$CFLAGS -O0 -g"
-  - make
+        LDFLAGS="$LDFLAGS" ${LDLIBS+LDLIBS="$LDLIBS"}
+  - grep -q freestanding-pass .config || make
         QMAKEFLAGS="$QMAKEFLAGS QMAKE_CXX=$CXX"
         tests/Makefile
-  - (cd tests && make check -k
+  - grep -q freestanding-pass .config ||
+        (cd tests && make check -k
         TESTRUNNER=`which valgrind 2>/dev/null`)
   - make -s clean
   - ./scripts/update-docs.sh
